+++ tests/ConnectionTest.php
@@ -10,7 +10,6 @@ use Yiisoft\Db\Driver\Pdo\PdoConnectionInterface;
 use Yiisoft\Db\Exception\Exception;
 use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
-use Yiisoft\Db\Mssql\Column\ColumnFactory;
 use Yiisoft\Db\Mssql\Tests\Support\TestTrait;
 use Yiisoft\Db\Tests\Common\CommonConnectionTest;
 use Yiisoft\Db\Transaction\TransactionInterface;
@@ -87,12 +86,5 @@ final class ConnectionTest extends CommonConnectionTest
         $db = $this->getConnection();
 
         $this->assertSame(PDO::ERRMODE_EXCEPTION, $db->getActivePDO()?->getAttribute(PDO::ATTR_ERRMODE));
-    }
-
-    public function testGetColumnFactory(): void
-    {
-        $db = $this->getConnection();
-
-        $this->assertInstanceOf(ColumnFactory::class, $db->getColumnFactory());
     }
 }